+++ NEWS.md
@@ -22,6 +22,7 @@ igraph objects or adjacency matrix
 * ```combineBCR()``` offers single/dual chain clustering, aa or nt sequences, adaptive filtering of V and J genes and normalized or straight edit distance calculations
 * ```percentGeneUsage()``` now is the underlying function for ```percentGenes()```, ```percentVJ()```, 
 and ```vizGenes()``` and allows for percent, proportion and raw count quantification.
+* Added common theme (internal ```.themeRepertoire()```) to all plots and allow users to pass arguments to it
 
 ## BUG FIXES
 * ```clonalCompare()``` issue with plotting a 0 row data frame now errors with message